fix: adapt web build to phones (#65)
* fix: better layout for phones

* refactor: continue refactoring debugger structure

* fix: allocate ram and hram in heap to avoid stack overflow in wasm

* refactor: better structure in debugger

* refactor: more refactor, but still needs work

* refactor: complete refactor, needs cleanup

* fix: fix loop of file dialog opening

* fix: return to html button to avoid 'VM595:1 File chooser dialog can only be shown with a user activation'

* feat: better buttons for phones

+++ core/src/ppu/mod.rs
@@ -207,7 +207,7 @@ impl Ppu {
     }
 
     #[inline(always)]
-    pub fn get_scroll(&self) -> (i32, i32) { (self.scroll_x as i32, self.scroll_y as i32) }
+    pub fn get_scroll(&self) -> (u8, u8) { (self.scroll_x, self.scroll_y) }
 
     #[inline(always)]
     fn set_mode(&mut self, mode: LCDMode) { self.lcd_status = (self.lcd_status & 0xFC) | mode as u8; }
